feat: add verification script
