+++ package.json
@@ -11,6 +11,10 @@
     "deploy:mainnet": "npx hardhat run scripts/1_deploy_contracts.js --network mainnet",
     "deploy:polygon": "npx hardhat run scripts/1_deploy_contracts.js --network polygon",
     "deploy:sepolia": "npx hardhat run scripts/1_deploy_contracts.js --network sepolia",
+    "verify:holesky": "npx hardhat run scripts/2_verify_contracts.js --network holesky",
+    "verify:mainnet": "npx hardhat run scripts/2_verify_contracts.js --network mainnet",
+    "verify:polygon": "npx hardhat run scripts/2_verify_contracts.js --network polygon",
+    "verify:sepolia": "npx hardhat run scripts/2_verify_contracts.js --network sepolia",
     "test": "npx hardhat test",
     "lint:js": "npx eslint .",
     "lint:js:fix": "npx eslint . --fix",
