+++ scripts/1_deploy_contracts.js
@@ -2,25 +2,10 @@ require('dotenv').config()
 const { ethers, upgrades } = require('hardhat')
 const {
   getAdminAddress,
+  getInitializerArguments,
   printBoxedTitle,
-  verifyImplementationContract,
-  verifyProxyAdminContract,
-  verifyProxyContract
+  verifyContract
 } = require('./utils')
-
-// Initialize contract arguments
-function getInitializerArguments (admin) {
-  return [
-    admin, // Admin address
-    'Test CMTA Token', // nameIrrevocable
-    'TCMTAT', // symbolIrrevocable
-    18, // decimalsIrrevocable
-    'TCMTAT_ISIN', // tokenId
-    'https://cmta.ch', // terms
-    'TCMTAT_info', // information
-    0 // flag
-  ]
-}
 
 // Deploy CMTAT_BASE Contract
 async function deployCMTATBase () {
@@ -36,34 +21,6 @@ async function deployCMTATBase () {
 
   console.log('Deployed contract object:\n', proxyContract)
   return proxyContract
-}
-
-// Main verification function
-async function verifyContract (proxyContract) {
-  const delay = 30000 // 30-second delay for network propagation
-  console.log(`\nWaiting ${delay / 1000} seconds for the network to propagate...`)
-  await new Promise(resolve => setTimeout(resolve, delay))
-
-  const constructorArguments = getInitializerArguments(await getAdminAddress())
-
-  // Helper function for individual verifications
-  async function performVerification (verificationFunction, contractType) {
-    try {
-      await verificationFunction()
-      console.log(`${contractType} contract verified successfully.`)
-    } catch (error) {
-      console.error(`Error verifying ${contractType} contract:`, error)
-    }
-  }
-
-  // Verify Proxy Contract
-  await performVerification(() => verifyProxyContract(proxyContract, constructorArguments), 'Proxy')
-
-  // Verify Proxy Admin Contract
-  await performVerification(() => verifyProxyAdminContract(proxyContract), 'Proxy Admin')
-
-  // Verify Implementation Contract
-  await performVerification(() => verifyImplementationContract(proxyContract), 'Implementation')
 }
 
 // Main function
